changed upsert logic for skills, renamed folder to cv_form
